+++ pmd/src/main/java/net/sourceforge/pmd/cpd/CPDConfiguration.java
@@ -60,7 +60,7 @@ public class CPDConfiguration extends AbstractConfiguration {
 	@Parameter(names = "--ignore-annotations", description = "Ignore language annotations when comparing text", required = false)
 	private boolean ignoreAnnotations;
 
-	@Parameter(names = "--skip-lexical-errors", description = "Skip files which can't be tokenized due to invalid characters", required = false)
+	@Parameter(names = "--skip-lexical-errors", description = "Skip files which can't be tokenized due to invalid characters instead of aborting CPD", required = false)
 	private boolean skipLexicalErrors = false;
 
 	@Parameter(names = "--files", variableArity = true, description = "List of files and directories to process", required = false)